backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
pydantic==2.10.4
openai>=1.30.0
python-dotenv==1.0.1
tavily-python>=0.3.0
